Special handling for when a union's discriminator field is an enum.
If a union uses a field as a discriminator but that field is defined as
an enum type, then we have to reference it by name differently. This is
because enum names have a different conversion from a QAPI name to a
Java name. Sometimes, this name cannot easily be derived from the known
information so far.

Concretely, the following bug introduced this:
1. The type "SysEmuTarget" is defined as an enum. One of its values is
   "x86_64".
2. The type "CpuInfoFast" uses a field of the above type as a
   discriminator.
3. In Java generated code, it was trying to reference SysEmuType.x8664.
   This type did not exist in the enum. It only has "x86_64".

Now we save all enum fields and check if a union is referencing an
enum explicitly. If it is, then we use the enum name for that field.